Extract ApproveButton from Swap component
